requirements.txt
# Core dependencies
python-telegram-bot==20.3
openai>=1.0.0
python-dotenv==1.0.0

# Web crawling (simplified)
requests==2.31.0
beautifulsoup4==4.12.3

# Video processing
moviepy==1.0.3
edge-tts==6.1.10

# HTTP requests
aiohttp==3.10.5

# Data processing
pandas==2.1.4
numpy==1.24.4

# Async support
asyncio-mqtt==0.16.1

# Logging
loguru==0.7.2

# Additional utilities
pydantic==2.5.3
python-dateutil==2.8.2
